backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
mypy-extensions==1.0.0
numpy==2.2.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
pymongo==4.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
